openai-agents-project/pyproject.toml
[project]
name = "openai-agents-project"
version = "0.1.0"
description = "OpenAI Agents SDK with multi-agent workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.4.2",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]
